variable fix, next message

+++ src/main/java/com/grouppy/spring/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable Long id) {
+    public User getUserById(@PathVariable int id) {
         return userRepository.findById(id).orElse(null);
     }
 }
