+++ src/main/java/com/grouppy/spring/UserRepository.java
@@ -2,4 +2,4 @@ package com.grouppy.spring;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //what does this do ?
-public interface UserRepository extends JpaRepository<User, Long> {}
+public interface UserRepository extends JpaRepository<User, Integer> {}
